+++ source/nuxtend/pages/portal/ui/meta/search/fullSearch/index.vue
@@ -1,92 +1,101 @@
-<template lang="html">
+<template>
   <div class="contents-wrap">
-    <h2 class="hidden">데이터 검색</h2>
+    <h2 class="hidden">데이터</h2>
     <div class="contents-subject">
       <group-breadcrumb></group-breadcrumb>
-      <div class="subject">
-        <h3 class="subject__title">데이터 통합검색</h3>
-        <p class="subject__desc">
-          데이터를 트리구조로 검색할 수 있는 트리뷰를 제공합니다.
-        </p>
-      </div>
+      <subject></subject>
     </div>
     <div class="content-top">
       <div class="contents-top__search">
-        <search-input-field
-          :searchKeyword="searchKeyword"
-          :use-recent-search="true"
-          :use-inner-search="true"
-          @search="searchBtnClick"
-          @filterCheck="rescanFilterCheck"
-        ></search-input-field>
+        <search-input-field></search-input-field>
         <div class="contents-top__recommend">
-          <h4>추천검색어</h4>
+          <!-- 추천검색어 : 검색어 입력창에 검색어 입력 없이 빈 상태일 때 추천 검색어 노출 -->
+          <p>추천검색어</p>
+
           <basic-tag-list
-            :tagList="tagList"
+            :tagList="recommendTagList"
             previousText="#"
             :useCancelButton="false"
             @tagClick="recommendTagClick"
           ></basic-tag-list>
+          <!-- //추천검색어 -->
         </div>
       </div>
-      <search-result-box
-        class="contents-search-result"
-        :showSearchResultBox="showSearchResultBox"
-        :searchResultSuccess="searchResultSuccess"
-      >
-        <template
-          v-if="searchKeywordList.length < 2"
-          v-slot:resultSuccessTrueText
+      <div class="contents-search-result">
+        <p class="contents-search-result__text">
+          <strong>SGI</strong>에 대한 검색결과, 총 <strong>150</strong> 건
+          입니다.
+        </p>
+      </div>
+      <div class="contents-top__detail">
+        <base-button
+          class="detail-button-default"
+          :class="isDetailOpen ? 'detail-button--open' : ''"
+          @click="toggleDetail"
+          title="상세검색"
+          >데이터 상세검색</base-button
         >
-          <strong>'{{ searchKeyword }}'</strong>에 대한 검색결과 총
-          <strong>{{ contents.totalcount }}</strong> 건이 검색되었습니다.
-        </template>
-
-        <!-- TODO: '교통체증' & '대중교통' & '자전거' 와 같이 item 사이에 구분자 css 처리 필요 / 임시로 일단 처리해둠 -->
-        <template v-else v-slot:resultSuccessTrueText>
-          <strong
-            v-for="(item, index) in searchKeywordList"
-            :key="'searchKeyword_' + index"
-            >'{{ item }}'
-            <span
-              v-if="searchKeywordList.length - 1 !== index"
-              style="color: black"
-              >&</span
-            >
-          </strong>
-          검색 결과, 총 <strong>{{ contents.totalcount }}</strong> 건 입니다.
-        </template>
-      </search-result-box>
-
-      <!-- TODO : 한자연 기획 변경 (상세검색 선택패널 삭제) 으로 삭제하였습니다.-->
-      <!--      <div class="contents-top__detail">-->
-      <!--        <base-button-->
-      <!--          class="detail-button-default"-->
-      <!--          :class="isDetailOpen ? 'detail-button-open' : ''"-->
-      <!--          @click="toggleDetail"-->
-      <!--          title="상세검색"-->
-      <!--          >상세검색</base-button-->
-      <!--        >-->
-      <!--      </div>-->
+        <div
+          class="contents-top__filter"
+          :class="isDetailOpen ? 'contents-top__filter--open' : ''"
+        >
+          <div class="filter-group">
+            <div class="filter-group__options">
+              <!-- 토글스위치 -->
+              <base-switch>
+                <span>전체 선택</span>
+              </base-switch>
+              <!-- //토글스위치 -->
+              <div class="filter-group__notice">
+                <p class="info-text">분류를 선택하세요. (중복가능)</p>
+              </div>
+            </div>
+            <!-- 필터그룹핑 -->
+            <div class="search-filter-group search-filter-group--horizontal">
+              <group-search-filter
+                :component-key="treeObj.componentKey"
+                :tree-obj="treeObj"
+                :tree-key="treeObj.treeKey"
+              ></group-search-filter>
+              <group-search-filter
+                :component-key="treeObj.componentKey"
+                :tree-obj="treeObj"
+                :tree-key="treeObj.treeKey"
+              ></group-search-filter>
+              <group-search-filter
+                :component-key="treeObj.componentKey"
+                :tree-obj="treeObj"
+                :tree-key="treeObj.treeKey"
+              ></group-search-filter>
+            </div>
+            <!-- //필터그룹핑 -->
+          </div>
+        </div>
+      </div>
     </div>
     <div class="contents">
       <!--사이드메뉴-->
-      <aside>
+      <aside class="aside">
         <h3 class="hidden">검색필터 상세선택</h3>
-        <!--        높이 임시로 넣었습니다.-->
-        <div class="aside__category">
-          <div
-            class="aside__category-default"
-            :class="isCategoryExpand ? 'aside__category-expand' : ''"
-          >
-            <h4 class="aside__title">카테고리</h4>
-            <base-button
-              class="aside__button-expand"
-              :class="isCategoryExpand ? 'aside__button-default' : ''"
-              @click="toggleCategoryExpand"
-            ></base-button>
-
-            <!-- 카테고리 트리 component -->
+        <div class="aside-category">
+          <div class="aside-category__options">
+            <!-- 토글스위치 -->
+            <base-switch>
+              <span>전체 선택</span>
+            </base-switch>
+            <!-- //토글스위치 -->
+          </div>
+          <div class="search-filter-group search-filter-group--vertical">
+            <group-search-filter
+              :component-key="treeObj.componentKey"
+              :tree-obj="treeObj"
+              :tree-key="treeObj.treeKey"
+            ></group-search-filter>
+            <group-search-filter
+              :component-key="treeObj.componentKey"
+              :tree-obj="treeObj"
+              :tree-key="treeObj.treeKey"
+            ></group-search-filter>
             <group-search-filter
               :component-key="treeObj.componentKey"
               :tree-obj="treeObj"
@@ -94,108 +103,128 @@
             ></group-search-filter>
           </div>
         </div>
-
-        <!-- filter checkbox component : 제공기관 -->
-        <checkbox-filter-list
-          :use-expand-button="true"
-          :filter-id="CONSTANTS.FILTER.KEYS.PROVIDER"
-          filter-title="제공기관"
-          :filter-list="searchFilterList[CONSTANTS.FILTER.KEYS.PROVIDER]"
-          :select-checkbox-list="
-            selectSearchFilterList[CONSTANTS.FILTER.KEYS.PROVIDER]
-          "
-          @changeCheckboxList="changeCheckboxList"
-        >
-        </checkbox-filter-list>
-
-        <!-- filter checkbox component : 데이터타입-->
-        <checkbox-filter-list
-          :use-expand-button="false"
-          :filter-id="CONSTANTS.FILTER.KEYS.TYPE"
-          filter-title="데이터타입"
-          :filter-list="searchFilterList[CONSTANTS.FILTER.KEYS.TYPE]"
-          :select-checkbox-list="
-            selectSearchFilterList[CONSTANTS.FILTER.KEYS.TYPE]
-          "
-          @changeCheckboxList="changeCheckboxList"
-        >
-        </checkbox-filter-list>
-
-        <!-- filter checkbox component : 키워드 -->
-        <checkbox-filter-list
-          :use-expand-button="false"
-          :filter-id="CONSTANTS.FILTER.KEYS.KEYWORD"
-          filter-title="키워드"
-          :filter-list="searchFilterList[CONSTANTS.FILTER.KEYS.KEYWORD]"
-          :select-checkbox-list="
-            selectSearchFilterList[CONSTANTS.FILTER.KEYS.KEYWORD]
-          "
-          @changeCheckboxList="changeCheckboxList"
-        >
-        </checkbox-filter-list>
       </aside>
-
       <!-- // 사이드메뉴 -->
       <section>
-        <!--탭 -->
-        <group-tab
-          :tabList="tabList"
-          :useTabNum="true"
-          @tabClick="tabClick"
-        ></group-tab>
-        <!-- // 탭 -->
-        <div class="contents__list">
-          <div class="contents__list-head">
-            <h4>
-              전체<span>{{ contents.totalcount }}</span>
-            </h4>
-            <div class="list-head__options">
-              <div class="list-head__options-sort">
-                <basic-option
-                  radioClass="radio--check"
-                  :optionList="sortOptionList"
-                  type="text"
-                  name="radio-sort"
-                  @checkOption="sortOptionChange"
-                ></basic-option>
+        <div class="contents__result">
+          <!--탭 -->
+          <div class="contents__tab">
+            <group-tab></group-tab>
+          </div>
+          <!-- // 탭 -->
+          <div class="contents__list">
+            <div class="contents__list_inner">
+              <!-- heading-group -->
+              <div class="heading-group">
+                <h4>전체<span>(<strong>152</strong>건)</span></h4>
+                <div class="heading-group__options">
+                  <!-- sort -->
+                  <div class="options-sort">
+                    <!-- radios -->
+                    <div class="radios">
+                      <span class="hidden">정렬선택</span>
+                      <base-radio
+                        name="radio-check"
+                        radioId="radio-sort1"
+                        class="radio--check"
+                        checked
+                      >
+                        <template v-slot:label
+                          ><span class="radio-check-first"
+                            >정확도순</span
+                          ></template
+                        >
+                      </base-radio>
+                      <base-radio
+                        name="radio-check"
+                        radioId="radio-sort2"
+                        class="radio--check"
+                      >
+                        <template v-slot:label
+                          ><span>최신등록순</span></template
+                        >
+                      </base-radio>
+                      <base-radio
+                        name="radio-check"
+                        radioId="radio-sort3"
+                        class="radio--check"
+                      >
+                        <template v-slot:label><span>조회순</span></template>
+                      </base-radio>
+                      <base-radio
+                        name="radio-check"
+                        radioId="radio-sort4"
+                        class="radio--check"
+                      >
+                        <template v-slot:label
+                          ><span>다운로드순</span></template
+                        >
+                      </base-radio>
+                    </div>
+                    <!-- select -->
+                    <base-select class="select w-107">
+                      <template v-slot:title>
+                        <span class="select-selector__title">20개씩</span>
+                      </template>
+                      <template v-slot:list>
+                        <li class="select-container__item">
+                          <button
+                            class="select-container__button"
+                            type="button"
+                            role="option"
+                          >
+                            <span class="select-container__text">12개씩</span>
+                          </button>
+                        </li>
+                        <li class="select-container__item">
+                          <button
+                            class="select-container__button"
+                            type="button"
+                            role="option"
+                          >
+                            <span class="select-container__text">20개씩</span>
+                          </button>
+                        </li>
+                        <li class="select-container__item">
+                          <button
+                            class="select-container__button"
+                            type="button"
+                            role="option"
+                          >
+                            <span class="select-container__text">40개씩</span>
+                          </button>
+                        </li>
+                      </template>
+                    </base-select>
+                  </div>
+                </div>
               </div>
-              <div class="list-head__options-view">
-                <basic-option
-                  class="radios--toggle"
-                  radioClass="radio--icon"
-                  :optionList="viewOptionList"
-                  type="icon"
-                  name="radio-view"
-                  @checkOption="viewOptionChange"
-                ></basic-option>
-              </div>
+              <!-- <div class="list-head__options-view">
+                <div class="radios radios--toggle">
+                  <base-radio name="radio-test-icon" radioId="radio1" class="radio--icon" checked>
+                    <template v-slot:label>
+                      <svg-icon name="list_katech" class="svg-icon" />
+                      <span class="hidden">목록형</span>
+                    </template>
+                  </base-radio>
+                  <base-radio name="radio-test-icon" radioId="radio2" class="radio--icon">
+                    <template v-slot:label>
+                      <svg-icon name="list_card_katech" class="svg-icon" />
+                      <span class="hidden">카드형</span>
+                    </template>
+                  </base-radio>
+                </div>
+              </div> 2022-06-16 DEL : KT는 카드형이 없음 -->
             </div>
           </div>
           <div>
-            <!--   카드형인 경우, data-box-list--card 클래스 추가-->
-            <search-list
-              :class="isListCard ? 'data-box-list--card' : ''"
-              :list="contents.searchList"
-              :myFavoriteDataList="myFavoriteDataList"
-              :searchKeyword="searchKeyword"
-              :searchKeywordList="searchKeywordList"
-              @dataBoxClick="listDataBoxClick"
-              @keywordClick="listKeywordClick"
-              @myFavoriteDataClick="myFavoriteDataClick"
-            ></search-list>
+            <!-- 검색된 목록 리스트 -->
+            <!-- TODO: 실제 search-list 연동 필요 -->
+            <!-- <search-list :list="contents"></search-list> -->
           </div>
         </div>
         <div class="contents__pagination">
-          <group-pagination
-            :paging-key="paginationKey"
-            :paging-object="{
-              [CONSTANTS.PAGING.ITEMS_PER_PAGE]: 5,
-              [CONSTANTS.PAGING.VISIBLE_PAGES]: 3,
-              [CONSTANTS.PAGING.PAGE]: 1
-            }"
-            @pagingEvent="getGridData"
-            :show-test-table="false"
-          ></group-pagination>
+          <group-pagination></group-pagination>
         </div>
       </section>
     </div>
@@ -203,70 +232,50 @@
 </template>
 
 <script type="text/javascript">
-import BaseButton from "@common/atoms/base-button/base-button";
-import BaseCheckbox from "@common/atoms/base-checkbox/base-checkbox";
+import BaseRadio from "@component/common/atoms/base-radio/base-radio";
+import BasicTagList from "@component/common/atoms/basic-tag-list/basic-tag-list";
+import BaseSelect from "@component/common/atoms/base-select/base-select";
+import BaseButton from "@component/common/atoms/base-button/base-button";
+import BaseCheckbox from "@component/common/atoms/base-checkbox/base-checkbox";
+import BaseSwitch from "@component/common/atoms/base-switch/base-switch";
 import GroupTab from "@component/common/molecules/group-tab/group-tab";
 import GroupPagination from "@component/common/molecules/group-pagination/group-pagination";
-import GroupBreadcrumb from "@common/molecules/group-breadcrumb/group-breadcrumb";
+import GroupBreadcrumb from "@component/common/molecules/group-breadcrumb/group-breadcrumb";
+import Subject from "@component/common/organisms/subject/subject.vue";
 import GroupSearchFilter from "@component/common/molecules/group-search-filter/group-search-filter";
 import SearchList from "@component/common/organisms/search-list/search-list.vue";
 import SearchInputField from "@component/common/organisms/search-input-field/search-input-field.vue";
-import SearchResultBox from "@component/common/atoms/search-result-box/search-result-box";
-import BasicTagList from "@component/common/atoms/basic-tag-list/basic-tag-list";
-import BasicOption from "@component/common/atoms/basic-option/basic-option";
-import { mapActions, mapGetters } from "vuex";
-import CheckboxFilterList from "@component/common/molecules/checkbox-filter-list/checkbox-filter-list.vue";
+import { mapGetters } from "vuex";
 
 export default {
-  name: "fullSearch",
+  name: "Index",
+  layout: "kt/kt",
+  async asyncData({ store }) {
+    await store.dispatch("kt/keyword-search/getContents");
+  },
   computed: {
     ...mapGetters({
-      contents: "meta/keyword-search/contents",
-      CONSTANTS: "defaults/constants/CONSTANTS",
-      keyword: "meta/keyword-search/searchKeyword"
-    }),
-    ...mapGetters("meta/search/search", [
-      "searchFilterList",
-      "selectSearchFilterList"
-    ])
+      contents: "kt/keyword-search/contents"
+    })
   },
   components: {
-    CheckboxFilterList,
+    BaseRadio,
     BaseButton,
+    BasicTagList,
+    BaseSelect,
     BaseCheckbox,
+    BaseSwitch,
     GroupPagination,
     GroupBreadcrumb,
+    Subject,
     GroupSearchFilter,
     GroupTab,
     SearchList,
-    SearchInputField,
-    SearchResultBox,
-    BasicTagList,
-    BasicOption
+    SearchInputField
   },
   data() {
     return {
-      isDetailOpen: true,
-      isCategoryExpand: false,
-      isListCard: false,
-      filterObj: {
-        category: {
-          label: "카테고리",
-          componentType: "checkbox"
-        },
-        provider: {
-          label: "제공기관",
-          componentType: "checkbox"
-        },
-        dataType: {
-          label: "데이터 타입",
-          componentType: "checkbox"
-        },
-        treeView: {
-          label: "트리뷰",
-          componentType: "tree"
-        }
-      },
+      isDetailOpen: false,
       treeObj: {
         componentKey: "metaTreeKey",
         treeRestApi: "/getCategoryList",
@@ -277,146 +286,27 @@ export default {
           PRNTS_ID: "prnts_id" // parent key
         }
       },
-      searchKeyword: "",
-      searchKeywordList: [],
-      rescanFilterChecked: false,
-      showSearchResultBox: false,
-      searchResultSuccess: false,
-      tagList: [
-        { itemId: 1, itemName: "결빙사고" },
-        { itemId: 2, itemName: "사고다발지역" },
-        { itemId: 3, itemName: "교통체증" },
-        { itemId: 4, itemName: "대중교통" },
-        { itemId: 5, itemName: "자전거" }
-      ],
-      tabList: [
-        { num: "999+", title: "전체" },
-        { num: "133", title: "내부데이터" },
-        { num: "0", title: "CKAN" },
-        { num: "71", title: "분원데이터" }
-      ],
-      paginationKey: "fullSearchPagination",
-      sortOptionList: [
-        { label: "정확도순", option: "accuracy" },
-        { label: "수정일순", option: "updateDate" },
-        { label: "조회순", option: "view" }
-      ],
-      viewOptionList: [
-        { label: "목록형", option: "list", svgIconName: "list_katech" },
-        { label: "카드형", option: "card", svgIconName: "list_card_katech" }
-      ],
-      myFavoriteDataList: ["1", "3", "4"]
+      recommendTagList: [
+        { itemId: 1, itemName: "데이터사업" },
+        { itemId: 2, itemName: "업종" },
+        { itemId: 3, itemName: "콜등급" },
+        { itemId: 4, itemName: "트래픽" },
+        { itemId: 5, itemName: "CU" },
+        { itemId: 5, itemName: "RTPO" }
+      ]
     };
   },
   methods: {
-    ...mapActions("meta/keyword-search", ["getContents", "setSearchKeyword"]),
-    ...mapActions("meta/search/search", [
-      "getSearchFilterList",
-      "changeSearchFilterList"
-    ]),
-    toggleCategoryExpand: function () {
-      this.isCategoryExpand = !this.isCategoryExpand;
-    },
-    searchBtnClick(inputData) {
-      this.searchKeyword = inputData.trim();
-      this.search();
-    },
-    searchResultBox(show, success) {
-      this.showSearchResultBox = show;
-      this.searchResultSuccess = success;
-    },
-    search() {
-      if (this.searchKeyword.trim() === "") {
-        this.searchResultBox(false, false);
-        alert("값을 입력해주세요.");
-        return;
-      } else {
-        this.searchResultBox(true, true);
-      }
-
-      if (this.rescanFilterChecked) {
-        if (this.searchKeywordList.length >= 3) {
-          alert("결과 내 재검색 3회이상으로 추가 검색이 불가능 합니다.");
-          this.searchKeyword = "";
-          return;
-        } else if (this.searchKeywordList.includes(this.searchKeyword)) {
-          alert("동일한 검색어 입력으로 추가 검색이 불가능 합니다.");
-          return;
-        }
-      } else {
-        this.searchKeywordList = [];
-      }
-
-      this.searchKeywordList.push(this.searchKeyword);
-
-      this.getGridData();
+    toggleDetail: function () {
+      this.isDetailOpen = !this.isDetailOpen;
     },
     recommendTagClick(tagObj) {
-      this.searchKeyword = tagObj.itemName;
-      this.search();
-    },
-    tabClick({ tabObj }) {
-      console.log(tabObj);
-    },
-    getGridData() {
-      this.getContents({
-        paginationKey: this.paginationKey,
-        searchKeywordList: this.searchKeywordList
-      });
-    },
-    sortOptionChange(option) {
-      console.log("sort option: " + option);
-    },
-    viewOptionChange(option) {
-      console.log("view option: " + option);
-      this.isListCard = option === "card";
-    },
-    listKeywordClick(tagObj) {
-      console.log(tagObj.itemId + ", " + tagObj.itemName);
-    },
-    listDataBoxClick({ postId }) {
-      this.$router.push({
-        path: "/portal/ui/meta/search/fullSearch/detail",
-        query: { postId }
-      });
-    },
-    myFavoriteDataClick({ postId, bool }) {
-      console.log(postId + ", " + bool);
-    },
-    changeCheckboxList({ checkboxKey, changeList }) {
-      this.changeSearchFilterList({
-        key: checkboxKey,
-        changeList: changeList
-      });
-    },
-    rescanFilterCheck({ bool }) {
-      this.rescanFilterChecked = bool;
-      if (bool === false && this.searchKeyword === "") {
-        this.searchKeywordList = [];
-        this.searchKeyword = "";
-        this.searchResultBox(false, false);
-        this.getGridData();
-      }
-    },
-    searchDetailKeyword() {
-      this.search();
-      this.setSearchKeyword("");
+      console.log(tagObj);
     }
-  },
-  mounted() {
-    if (this.keyword === "") {
-      this.getGridData();
-    } else {
-      this.searchKeyword = this.keyword;
-      this.searchDetailKeyword();
-    }
-  },
-  created() {
-    this.getSearchFilterList();
   }
 };
 </script>
 
 <style lang="scss" scoped>
-@import "pages/portal/ui/meta/search/fullSearch/index";
+@import "./index.scss";
 </style>
